ai_models/split_merge/split.py
import os
from ai_models.split_merge.dataset.dataset import ImageDataset
from ai_models.split_merge.modules.split_modules import SplitModel
import json
from PIL import Image
import torch
from torchsummary import summary
import numpy as np
import cv2
import json

os.environ['CUDA_VISIBLE_DEVICES'] = '0'