requirements.txt
Pillow
streamlit
numpy
pdfplumber
streamlit_option_menu
opencv-python-headless
easyocr
streamlit-image-select
wired_table_rec
lineless_table_rec
table_cls
rapid_table
rapid_table[torch]
rapidocr_onnxruntime
onnxruntime-gpu
matplotlib
transformers
psutil
pytesseract
opencv-python
timm
accelerate>=0.20.3
pillow
pandas
google-cloud-aiplatform
apted
distance
lxml